Move RSRC class
We moved the RSRC class into its own file.
