+++ src/debloat/processor.py
@@ -17,39 +17,15 @@ import zlib
 from pefile import Structure, SectionStructure, DIRECTORY_ENTRY
 from typing import Generator, Iterable, Optional
 
+import debloat.utilities.rsrc as rsrc
+
 _KB = 1000
 _MB = _KB * _KB
 
 PACKER = {
     1 : "Nullsoft"
 }
-import enum
 
-class RSRC(enum.IntEnum):
-    CURSOR        = 0x01  # noqa
-    BITMAP        = 0x02  # noqa
-    ICON          = 0x03  # noqa
-    MENU          = 0x04  # noqa
-    DIALOG        = 0x05  # noqa
-    STRING        = 0x06  # noqa
-    FONTDIR       = 0x07  # noqa
-    FONT          = 0x08  # noqa
-    ACCELERATOR   = 0x09  # noqa
-    RCDATA        = 0x0A  # noqa
-    MESSAGETABLE  = 0x0B  # noqa
-    ICON_GROUP    = 0x0E  # noqa
-    VERSION       = 0x10  # noqa
-    DLGINCLUDE    = 0x11  # noqa
-    PLUGPLAY      = 0x13  # noqa
-    VXD           = 0x14  # noqa
-    ANICURSOR     = 0x15  # noqa
-    ANIICON       = 0x16  # noqa
-    HTML          = 0x17  # noqa
-    MANIFEST      = 0x18  # noqa
-
-    def __str__(self):
-        return self.name
-        
 def readable_size(value: int) -> str:
     '''Return bytes in human readable format.'''
     if value <= 1024:
@@ -226,9 +202,8 @@ def adjust_offsets(pe: pefile.PE, gap_offset: int, gap_size: int):
     return pe
     
 
-
 def refinery_strip(pe: pefile.PE, data: memoryview, block_size=_MB) -> int:
-    threshold = 2
+    threshold = 1
     alignment = pe.OPTIONAL_HEADER.FileAlignment
     data_overhang = len(data) % alignment
     result = data_overhang
@@ -280,8 +255,8 @@ def refinery_trim_resources(pe: pefile.PE, pe_data: bytearray) -> int:
             name = getattr(entry, 'name')
             numeric_id = getattr(entry, 'id')
             if not name:
-                if level == 0 and numeric_id in iter(RSRC):
-                    name = RSRC(entry.id)
+                if level == 0 and numeric_id in iter(rsrc.RSRC):
+                    name = rsrc.RSRC(entry.id)
                 elif numeric_id is not None:
                     name = str(numeric_id)
             name = name and str(name) or '?'
@@ -517,7 +492,8 @@ If file is a Nullsoft executable, but was not detected, the original file can
 be unpacked with the tool "UniExtract2".
                     """)
                     last_section = find_last_section(pe)
-                    end_of_real_data = last_section.PointerToRawData + last_section.SizeOfRawData 
+                    end_of_real_data = last_section.PointerToRawData + last_section.SizeOfRawData
+                    pe_data = pe_data[:end_of_real_data] 
                 else:
                     log_message("""
 Overlay was unable to be trimmed. Try unpacking with UniExtract2 or re-running
